+++ Sources/ALTENNetwork/NetworkSession.swift
@@ -36,12 +36,13 @@ public protocol NetworkSession: Actor {
     ///  - Returns: `NetworkSessionInterceptionResult` que indica si se debe continuar con el flujo o si se debe reintentar la petición con una nueva `URLRequest`
     func interceptResponse(networkSession: NetworkSession, originalRequest: URLRequest, retryNumber: Int, result: Result<NetworkSessionInterception, Error>) async throws -> NetworkSessionInterceptionResult
     
-    /// Este método se llamará antes de que comience cualquier request. Tiene como parámetro de entrada la URLRequest a la que se está llamando y su único objetivo es que sirva de caracter informativo.
+    /// Este método se llamará antes de que comience cualquier request. Tiene como parámetro de entrada la URLRequest a la que se está llamando. Este método tiene como objetivo modificar la Request antes de que se realice la petición.
     /// La implementación por defecto imprime el curl de la request sólo en entornos de debug a través de la condificón `#if DEBUG`
     /// - Parameters:
     /// - networkSession: `NetworkSession` que realiza la petición
     /// - originalRequest: `URLRequest` original a la que se está llamando
-    func requestStart(networkSession: NetworkSession, originalRequest: URLRequest) async
+    /// - Returns: `URLRequest` que se debe realizar
+    func interceptRequest(networkSession: NetworkSession, originalRequest: URLRequest) async -> URLRequestConvertible
     
     // Descarga el contenido de un `URLRequestConvertible` y lo almacena en memoria. `URLRequestConvertible` es en esencia un `URLRequest`. De forma básica podemos usar un `URL` o un `URLRequest` para realizar la petición
     /// - Parameters:
@@ -167,10 +168,11 @@ public protocol NetworkSession: Actor {
 
 extension NetworkSession {
     
-    public func requestStart(networkSession: NetworkSession, originalRequest: URLRequest) {
+    public func interceptRequest(networkSession: NetworkSession, originalRequest: URLRequest) async -> URLRequest {
 #if DEBUG
         print("[NetworkSession] - Start Request: \(originalRequest.curl)")
 #endif
+        return originalRequest
     }
     
     public func interceptResponse(networkSession: NetworkSession, originalRequest: URLRequest, retryNumber: Int, result: Result<NetworkSessionInterception, Error>) async throws -> NetworkSessionInterceptionResult {
@@ -212,13 +214,11 @@ extension NetworkSession {
     }
     
     private func _requestData(for request: URLRequestConvertible, delegate: URLSessionTaskDelegate?, retryNumber: Int = 0) async throws -> NetworkDataResponse {
-        let originalRequest = request.asURLRequest()
-        let urlRequest = request.asURLRequest()
-        await requestStart(networkSession: self, originalRequest: urlRequest)
+        let urlRequest = await interceptRequest(networkSession: self, originalRequest: request.asURLRequest()).asURLRequest()
         do {
             let response: NetworkDataResponse
             if #available(iOS 15, tvOS 15, *) {
-                response = try await NetworkDataResponse(dataResponse: session.data(for: urlRequest, delegate: delegate), originalRequest: originalRequest)
+                response = try await NetworkDataResponse(dataResponse: session.data(for: urlRequest, delegate: delegate), originalRequest: urlRequest)
             } else {
                 response = try await withCheckedThrowingContinuation { continuation in
                     let dataTask = session.dataTask(with: urlRequest) { data, response, error in
@@ -227,7 +227,7 @@ extension NetworkSession {
                             if let error = error {
                                 continuation.resume(throwing: error)
                             } else if let data = data, let response = response {
-                                continuation.resume(returning: NetworkDataResponse(dataResponse: (data, response), originalRequest: originalRequest))
+                                continuation.resume(returning: NetworkDataResponse(dataResponse: (data, response), originalRequest: urlRequest))
                             } else {
                                 continuation.resume(throwing: NetworkError.unknown)
                             }
@@ -238,9 +238,9 @@ extension NetworkSession {
                     dataTask.resume()
                 }
             }
-            return try await _handleResponseDataInterpection(result: .success(response), request: originalRequest, delegate: delegate, retryNumber: retryNumber)
+            return try await _handleResponseDataInterpection(result: .success(response), request: urlRequest, delegate: delegate, retryNumber: retryNumber)
         } catch {
-            return try await _handleResponseDataInterpection(result: .failure(error), request: originalRequest, delegate: delegate, retryNumber: retryNumber)
+            return try await _handleResponseDataInterpection(result: .failure(error), request: urlRequest, delegate: delegate, retryNumber: retryNumber)
         }
     }
     
@@ -289,13 +289,11 @@ extension NetworkSession {
     }
     
     private func _requestDownload(for request: URLRequestConvertible, delegate: URLSessionTaskDelegate?, retryNumber: Int = 0) async throws -> NetworkDownloadResponse {
-        let originalRequest = request.asURLRequest()
-        let urlRequest = request.asURLRequest()
-        await requestStart(networkSession: self, originalRequest: urlRequest)
+        let urlRequest = await interceptRequest(networkSession: self, originalRequest: request.asURLRequest()).asURLRequest()
         do {
             let response: NetworkDownloadResponse
             if #available(iOS 15, tvOS 15, *) {
-                response = try await NetworkDownloadResponse(dataResponse: session.download(for: urlRequest, delegate: delegate), originalRequest: originalRequest)
+                response = try await NetworkDownloadResponse(dataResponse: session.download(for: urlRequest, delegate: delegate), originalRequest: urlRequest)
             } else {
                 response = try await withCheckedThrowingContinuation { continuation in
                     let dataTask = session.downloadTask(with: urlRequest) { url, response, error in
@@ -304,7 +302,7 @@ extension NetworkSession {
                             if let error = error {
                                 continuation.resume(throwing: error)
                             } else if let url = url, let response = response {
-                                continuation.resume(returning: NetworkDownloadResponse(dataResponse: (url, response), originalRequest: originalRequest))
+                                continuation.resume(returning: NetworkDownloadResponse(dataResponse: (url, response), originalRequest: urlRequest))
                             } else {
                                 continuation.resume(throwing: NetworkError.unknown)
                             }
@@ -316,9 +314,9 @@ extension NetworkSession {
                     dataTask.resume()
                 }
             }
-            return try await _handleResponseDownloadInterpection(result: .success(response), request: originalRequest, delegate: delegate, retryNumber: retryNumber)
+            return try await _handleResponseDownloadInterpection(result: .success(response), request: urlRequest, delegate: delegate, retryNumber: retryNumber)
         } catch {
-            return try await _handleResponseDownloadInterpection(result: .failure(error), request: originalRequest, delegate: delegate, retryNumber: retryNumber)
+            return try await _handleResponseDownloadInterpection(result: .failure(error), request: urlRequest, delegate: delegate, retryNumber: retryNumber)
         }
     }
     
@@ -367,13 +365,11 @@ extension NetworkSession {
     }
     
     private func _requestUpload(for request: URLRequestConvertible, from bodyData: Data, delegate: URLSessionTaskDelegate?, retryNumber: Int = 0) async throws -> NetworkUploadResponse {
-        let originalRequest = request.asURLRequest()
-        let urlRequest = request.asURLRequest()
-        await requestStart(networkSession: self, originalRequest: urlRequest)
+        let urlRequest = await interceptRequest(networkSession: self, originalRequest: request.asURLRequest()).asURLRequest()
         do {
             let response: NetworkUploadResponse
             if #available(iOS 15, tvOS 15, *) {
-                response = try await NetworkUploadResponse(dataResponse: session.upload(for: urlRequest, from: bodyData, delegate: delegate), originalRequest: originalRequest, uploadType: .data(bodyData))
+                response = try await NetworkUploadResponse(dataResponse: session.upload(for: urlRequest, from: bodyData, delegate: delegate), originalRequest: urlRequest, uploadType: .data(bodyData))
             } else {
                 response = try await withCheckedThrowingContinuation { continuation in
                     let dataTask = session.uploadTask(with: urlRequest, from: bodyData) { data, response, error in
@@ -382,7 +378,7 @@ extension NetworkSession {
                             if let error = error {
                                 continuation.resume(throwing: error)
                             } else if let data = data, let response = response {
-                                continuation.resume(returning: NetworkUploadResponse(dataResponse: (data, response), originalRequest: originalRequest, uploadType: .data(bodyData)))
+                                continuation.resume(returning: NetworkUploadResponse(dataResponse: (data, response), originalRequest: urlRequest, uploadType: .data(bodyData)))
                             } else {
                                 continuation.resume(throwing: NetworkError.unknown)
                             }
@@ -393,9 +389,9 @@ extension NetworkSession {
                     dataTask.resume()
                 }
             }
-            return try await _handleResponseUploadInterpection(result: .success(response), request: originalRequest, from: bodyData, delegate: delegate, retryNumber: retryNumber)
+            return try await _handleResponseUploadInterpection(result: .success(response), request: urlRequest, from: bodyData, delegate: delegate, retryNumber: retryNumber)
         } catch {
-            return try await _handleResponseUploadInterpection(result: .failure(error), request: originalRequest, from: bodyData, delegate: delegate, retryNumber: retryNumber)
+            return try await _handleResponseUploadInterpection(result: .failure(error), request: urlRequest, from: bodyData, delegate: delegate, retryNumber: retryNumber)
         }
     }
     
@@ -444,13 +440,11 @@ extension NetworkSession {
     }
     
     private func _requestUpload(for request: URLRequestConvertible, fromFile fileURL: URL, delegate: URLSessionTaskDelegate?, retryNumber: Int = 0) async throws -> NetworkUploadResponse {
-        let originalRequest = request.asURLRequest()
-        let urlRequest = request.asURLRequest()
-        await requestStart(networkSession: self, originalRequest: urlRequest)
+        let urlRequest = await interceptRequest(networkSession: self, originalRequest: request.asURLRequest()).asURLRequest()
         do {
             let response: NetworkUploadResponse
             if #available(iOS 15, tvOS 15, *) {
-                response = try await NetworkUploadResponse(dataResponse: session.upload(for: urlRequest, fromFile: fileURL, delegate: delegate), originalRequest: originalRequest, uploadType: .file(fileURL))
+                response = try await NetworkUploadResponse(dataResponse: session.upload(for: urlRequest, fromFile: fileURL, delegate: delegate), originalRequest: urlRequest, uploadType: .file(fileURL))
             } else {
                 response = try await withCheckedThrowingContinuation { continuation in
                     let dataTask = session.uploadTask(with: urlRequest, fromFile: fileURL) { data, response, error in
@@ -459,7 +453,7 @@ extension NetworkSession {
                             if let error = error {
                                 continuation.resume(throwing: error)
                             } else if let data = data, let response = response {
-                                continuation.resume(returning: NetworkUploadResponse(dataResponse: (data, response), originalRequest: originalRequest, uploadType: .file(fileURL)))
+                                continuation.resume(returning: NetworkUploadResponse(dataResponse: (data, response), originalRequest: urlRequest, uploadType: .file(fileURL)))
                             } else {
                                 continuation.resume(throwing: NetworkError.unknown)
                             }
@@ -470,9 +464,9 @@ extension NetworkSession {
                     dataTask.resume()
                 }
             }
-            return try await _handleResponseUploadInterpection(result: .success(response), request: originalRequest, fromFile: fileURL, delegate: delegate, retryNumber: retryNumber)
+            return try await _handleResponseUploadInterpection(result: .success(response), request: urlRequest, fromFile: fileURL, delegate: delegate, retryNumber: retryNumber)
         } catch {
-            return try await _handleResponseUploadInterpection(result: .failure(error), request: originalRequest, fromFile: fileURL, delegate: delegate, retryNumber: retryNumber)
+            return try await _handleResponseUploadInterpection(result: .failure(error), request: urlRequest, fromFile: fileURL, delegate: delegate, retryNumber: retryNumber)
         }
     }